requirements.txt
pyyaml>=6.0
fastapi>=0.110
uvicorn>=0.30
google-genai>=1.0
python-dotenv>=1.0
httpx>=0.27
edge-tts>=6.1
faster-whisper>=1.0
pipecat-ai>=0.0.1
openai>=1.0
anthropic>=0.40
psycopg2-binary>=2.9
sqlalchemy>=2.0
alembic>=1.13
passlib>=1.7
bcrypt>=4.1
python-jose>=3.3
python-multipart>=0.0.9
pydantic>=2.0
pydantic-settings>=2.0
flask>=3.0
websockets>=12.0
